+++ node/chain/shard/block_generator.go
@@ -11,12 +11,10 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"strconv"
 	"time"
 
 	mmtree "gitlab.com/jaxnet/core/merged-mining-tree"
 	"gitlab.com/jaxnet/jaxnetd/database"
-	"gitlab.com/jaxnet/jaxnetd/jaxutil"
 	"gitlab.com/jaxnet/jaxnetd/node/mining"
 	"gitlab.com/jaxnet/jaxnetd/txscript"
 	"gitlab.com/jaxnet/jaxnetd/types"
@@ -157,22 +155,7 @@ func (c *BlockGenerator) AcceptBlock(blockHeader wire.BlockHeader) error {
 }
 
 func (c *BlockGenerator) CalcBlockSubsidy(height int32, header wire.BlockHeader) int64 {
-	switch pow.ShardEpoch(height) {
-	case 1:
-		// (Di * Ki) * jaxutil.SatoshiPerJAXCoin
-		d := pow.GetDifficulty(header.Bits())
-		k1 := pow.UnpackRat(header.K())
-		rewardStr := new(big.Rat).Mul(d, k1).FloatString(4)
-		reward, err := strconv.ParseFloat(rewardStr, 64)
-		if err != nil {
-			return (20 * jaxutil.SatoshiPerJAXCoin) >> uint(height/210000)
-		}
-		return int64(reward * jaxutil.SatoshiPerJAXCoin)
-
-	default:
-		// Equivalent to: baseSubsidy / 2^(height/subsidyHalvingInterval)
-		return (20 * jaxutil.SatoshiPerJAXCoin) >> uint(height/210000)
-	}
+	return pow.CalcShardBlockSubsidy(height, header.Bits(), header.K())
 }
 
 func (c *BlockGenerator) generateBeaconHeader(nonce uint32, burnReward int) (*wire.BeaconHeader, wire.CoinbaseAux, error) {